+++ Installer/Provider/DatabaseSetupCommandsProvider.php
@@ -1,4 +1,4 @@
-<?php namespace VS\ApplicationInstalatorBundle\Installer\Provider;
+<?php namespace Vankosoft\ApplicationInstalatorBundle\Installer\Provider;
 
 use Doctrine\Bundle\DoctrineBundle\Registry;
 use Doctrine\DBAL\Schema\AbstractSchemaManager;