+++ CHANGELOG.md
@@ -24,7 +24,7 @@
 ============================================
 * New Features:
   - Update Application configs package.
-  - Update application Kernel Template to sextend VS\ApplicationBundle\Component\Application\Kernel.
+  - Update application Kernel Template to sextend Vankosoft\ApplicationBundle\Component\Application\Kernel.
 
 
 1.3.6	|	Release date: **17.10.2021**